requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
datetime-truncate==1.1.1
dj-database-url==1.2.0
Django==4.1.7
django-admin-charts==1.3.0
django-admin-soft-dashboard==1.0.7
django-bower==5.2.0
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
djangorestframework==3.14.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
# mysqlclient==2.1.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.14
